requirements.txt
asgiref==3.8.1
attrs==24.3.0
Django==4.2.17
django-cors-headers==4.2.0
django-environ==0.10.0
django-filter==24.3
django-leaflet==0.29.1
djangorestframework==3.14.0
djangorestframework-gis==1.1
drf-spectacular==0.26.5
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
psycopg==3.1.20
psycopg-binary==3.1.20
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.22.3
sqlparse==0.5.3
typing_extensions==4.12.2
uritemplate==4.1.1